Move to XML-based messaging via sendPluginCommand
Player has been redesigned to use XML as messaging protocol (fixes #6).
The columns of tvPlayerList are no longer tied to the message string.

The XML messages are transmitted via sendPluginCommand (fixes #4).

Sending messages to channel's chat temporarily disabled, will return
as optional feature in one of the next commits.

Extensive refactoring of DayzServerIp. tvRemoteInfo is now
tvPlayerList, tbLocalInfo is now tbPlayer. Various variables and
functions have been renamed accordingly.

The debug button and dialog have been removed (fixes #8). With XML,
there is no easy way to use this anymore and its sole purpose was
to make debugging quick and easy for me. Since I haven't been using
it for a while, it's not a big loss.